requirements.txt
arxiv
openai
notion-client<3.0.0
python-dotenv
requests
